Remove unused tests, add service to controller
